requirements.txt
torch
torchvision
opencv-python
facenet-pytorch
numpy
matplotlib
scikit-learn
tqdm
pandas
streamlit
Pillow